# 2025-12-05 Start using Google Ai Ultra Plan.
feat: Implement dynamic image resizing, integrate OMR layout and answer key APIs, and enhance header image viewer with overlay logic and new documentation.

- Static JSON fetch for layout and key answers once when page loads. Any sheet navigation does not require to re-load answer-key. TODO: Check if change of class will use correct key.

- first request 750k, next request upon next record around 122kb.

- No Correct or Wrong icon, maybe not needed. Green color over student marked = correct other was incorrect. Less messy.

Key Answer is fetch with:
http://gt-omr-api-1.gt:8000/sheets/answer-key/12102011
Maybe not correct yet. But leave it as it for now as it works but probably not a good design according to domain based. Maybe should live within task itself so thatfuture need of 10 different type of exam key answers can be handled properly.